Added footer and clickhandlers for mobile
Added a clickhandler to handle the hover events that occur on desktop.
When you click on the about or contact sections in mobile, you click on the
new element that appears to close it
Created a footer with contact and copyright information as well as credits to
developers. Optimized it for smaller formats, still need to do midrange