Fixed various Bugs
Fixed a typo in the services panel on the middle column
Since the carousel was non functioning in safari, added code that sets
the carousel to the mobile slide show when in safari.
Also extended the width of the `what-we-buy-mobile` containter so it wasn't
showing the edges of the slides.